Cleans up Core.cpp includes and moves signal handling to Debug.(h/cpp)
