+++ library/include/Debug.h
@@ -28,7 +28,19 @@ redistribute it freely, subject to the following restrictions:
 #include <atomic>
 #include "Core.h"
 
+#ifdef _WIN32
+#include <windows.h>
+#include <DbgHelp.h>
+#pragma comment(lib, "dbghelp.lib")
+#else
+#include <execinfo.h>
+#endif
+
 namespace DFHack {
+
+void signal_handler(int sig);
+
+void install_signal_handler();
 
 /*! \file Debug.h
  * Light weight wrappers for runtime debug output filtering. The idea is to add
